move the new-computed polyfill into a util folder and move the supportsGetterSetter helper into the same file as the computed polyfill
