--- packages/model-fragments/lib/addons/ember-new-computed/utils/can-use-new-syntax.js
@@ -1,14 +1,0 @@
-var Ember = window.Ember;
-var supportsSetterGetter;
-
-try {
-  Ember.computed({
-    set: function() { },
-    get: function() { }
-  });
-  supportsSetterGetter = true;
-} catch(e) {
-  supportsSetterGetter = false;
-}
-
-export default supportsSetterGetter;
